requirements.txt
Flask
pytesseract
Pillow
numpy
faiss-cpu
sentence-transformers
opencv-python
spacy
langdetect
scikit-learn